test(experiments): the dataset suite is its own file, over a shared boundary
The dispatch-guard file held a third suite about which dataset a run resolves
its inputs from, which is a different subject from the file's name and put it
over the 300 source-line limit for a new file.

That suite is now `activeDatasetMappings.integration.test.ts`, and what the
two files share is `dispatchBoundary.ts`: the scripted nlpgo boundary, the
dataset columns and the reset. A `vi.mock` factory is hoisted per module and
cannot be imported, so each file still declares its own three, but they
delegate to the shared boundary rather than repeat it.

The boundary file imports no module that reaches the mocked post-event
module. It did at first, and the run then hung with no test output at all,
because a mock factory awaited an import that led back to the module it was
mocking. The file says so, since the next reader cannot guess it.
