+++ platform/app/src/server/experiments-v3/execution/__tests__/evaluatorDispatchGuard.integration.test.ts
@@ -3,48 +3,33 @@
  *
  * The nlpgo dispatch boundary (and env injection / dataset inlining) is mocked
  * and fed scripted server events, so these run the dispatch decision itself
- * without a live NLP service: whether the evaluator is sent at all, and which
- * dataset the inputs were resolved from.
+ * without a live NLP service: whether the evaluator is sent at all.
  */
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import type { EvaluatorConfig } from "~/experiments-v3/types";
 import type { StudioServerEvent } from "~/optimization_studio/types/events";
 
-const scripted = vi.hoisted(() => ({
-  component: [] as StudioServerEvent[],
-  dispatched: [] as Array<{ type: string; payload: Record<string, any> }>,
-}));
-
-vi.mock("~/app/api/workflows/post_event/post-event", () => ({
+vi.mock("~/app/api/workflows/post_event/post-event", async () => ({
   studioBackendPostEvent: vi.fn(
-    async ({
-      message,
-      onEvent,
-    }: {
-      message: { type: string; payload: Record<string, any> };
-      onEvent: (event: StudioServerEvent) => void;
-    }) => {
-      scripted.dispatched.push(message);
-      for (const event of scripted.component) onEvent(event);
-    },
+    (await import("./dispatchBoundary")).postEventToScript,
   ),
 }));
-vi.mock("~/optimization_studio/server/addEnvs", () => ({
-  addEnvs: vi.fn(async (event: unknown) => event),
+vi.mock("~/optimization_studio/server/addEnvs", async () => ({
+  addEnvs: vi.fn((await import("./dispatchBoundary")).leaveEventAsItIs),
 }));
-vi.mock("~/optimization_studio/server/loadDatasets", () => ({
-  loadDatasets: vi.fn(async (event: unknown) => event),
+vi.mock("~/optimization_studio/server/loadDatasets", async () => ({
+  loadDatasets: vi.fn((await import("./dispatchBoundary")).leaveEventAsItIs),
 }));
 
-import { executeCell, generateCells } from "../orchestrator";
+import { executeCell } from "../orchestrator";
 import type { EvaluationV3Event, ExecutionCell } from "../types";
+import {
+  datasetColumns,
+  evaluatorDispatches,
+  resetBoundary,
+  scripted,
+} from "./dispatchBoundary";
 
-const datasetColumns = [
-  { id: "question", name: "question", type: "string" },
-  { id: "expected", name: "expected", type: "string" },
-];
-
-/** An exact-match chip whose fields are mapped, or not mapped at all. */
 const gradingEvaluator = (isMapped: boolean): EvaluatorConfig => ({
   id: "eval-1",
   evaluatorType: "langevals/exact_match",
@@ -105,7 +90,8 @@ const makeCell = (evaluator: EvaluatorConfig): ExecutionCell => ({
   precomputedTargetOutput: { output: "yes" },
 });
 
-const run = async (cell: ExecutionCell): Promise<EvaluationV3Event[]> => {
+/** Run one cell to its end and collect every event it produced. */
+const runCell = async (cell: ExecutionCell): Promise<EvaluationV3Event[]> => {
   const events: EvaluationV3Event[] = [];
   for await (const event of executeCell(cell, "p1", datasetColumns, {})) {
     events.push(event);
@@ -113,19 +99,13 @@ const run = async (cell: ExecutionCell): Promise<EvaluationV3Event[]> => {
   return events;
 };
 
-const evaluatorDispatches = () =>
-  scripted.dispatched.filter((message) => message.type === "execute_component");
-
-beforeEach(() => {
-  scripted.component = [];
-  scripted.dispatched = [];
-});
+beforeEach(resetBoundary);
 
 describe("given an evaluator attached to a target column", () => {
   describe("when none of its fields are mapped", () => {
     /** @scenario "An evaluator with no resolved inputs reports an error instead of a pass" */
     it("reports the row as an error and never calls the evaluator", async () => {
-      const events = await run(makeCell(gradingEvaluator(false)));
+      const events = await runCell(makeCell(gradingEvaluator(false)));
 
       const results = events.filter((e) => e.type === "evaluator_result");
       expect(results).toHaveLength(1);
@@ -138,7 +118,7 @@ describe("given an evaluator attached to a target column", () => {
 
     /** @scenario "An evaluator with no resolved inputs reports an error instead of a pass" */
     it("names the evaluator and the fix in the row's details", async () => {
-      const events = await run(makeCell(gradingEvaluator(false)));
+      const events = await runCell(makeCell(gradingEvaluator(false)));
 
       const result = events.find((e) => e.type === "evaluator_result");
       expect((result?.result as { details?: string }).details).toContain(
@@ -166,7 +146,7 @@ describe("given an evaluator attached to a target column", () => {
         },
       ] as unknown as StudioServerEvent[];
 
-      const events = await run(makeCell(gradingEvaluator(true)));
+      const events = await runCell(makeCell(gradingEvaluator(true)));
 
       expect(evaluatorDispatches()).toHaveLength(1);
       expect(evaluatorDispatches()[0]?.payload.inputs).toEqual({
@@ -293,108 +273,6 @@ describe("given an evaluator run as its own column", () => {
       });
       const result = events.find((e) => e.type === "target_result");
       expect(result?.error).toBeUndefined();
-    });
-  });
-});
-
-describe("given a workbench with two datasets", () => {
-  describe("when the active dataset is not the first one", () => {
-    /** @scenario "The run reads its mappings from the dataset the rows come from" */
-    it("dispatches the evaluator with inputs from the active dataset", async () => {
-      const state = {
-        datasets: [
-          { id: "dataset-old", name: "Old" },
-          { id: "dataset-active", name: "Active" },
-        ],
-        activeDatasetId: "dataset-active",
-        targets: [
-          {
-            id: "target-1",
-            type: "prompt",
-            inputs: [{ identifier: "input", type: "str" }],
-            outputs: [{ identifier: "output", type: "str" }],
-            mappings: {
-              "dataset-active": {
-                input: {
-                  type: "source",
-                  source: "dataset",
-                  sourceId: "dataset-active",
-                  sourceField: "question",
-                },
-              },
-            },
-            localPromptConfig: {
-              llm: { model: "openai/gpt-5-mini", temperature: 0 },
-              messages: [{ role: "user", content: "{{input}}" }],
-              inputs: [{ identifier: "input", type: "str" }],
-              outputs: [{ identifier: "output", type: "str" }],
-            },
-          },
-        ],
-        evaluators: [
-          {
-            id: "eval-1",
-            evaluatorType: "langevals/exact_match",
-            dbEvaluatorId: "db-eval-1",
-            inputs: [
-              { identifier: "output", type: "str" },
-              { identifier: "expected_output", type: "str" },
-            ],
-            mappings: {
-              "dataset-active": {
-                "target-1": {
-                  output: {
-                    type: "source",
-                    source: "target",
-                    sourceId: "target-1",
-                    sourceField: "output",
-                  },
-                  expected_output: {
-                    type: "source",
-                    source: "dataset",
-                    sourceId: "dataset-active",
-                    sourceField: "expected",
-                  },
-                },
-              },
-            },
-          },
-        ],
-      } as unknown as Parameters<typeof generateCells>[0];
-
-      const cells = generateCells(
-        state,
-        [{ question: "is a dog an animal?", expected: "yes" }],
-        { type: "full" },
-      );
-      const cell = {
-        ...cells[0]!,
-        skipTarget: true,
-        precomputedTargetOutput: { output: "yes" },
-      };
-
-      scripted.component = [
-        {
-          type: "component_state_change",
-          payload: {
-            component_id: "target-1.eval-1",
-            execution_state: {
-              status: "success",
-              outputs: { score: 1, passed: true },
-            },
-          },
-        },
-      ] as unknown as StudioServerEvent[];
-
-      const events = await run(cell);
-
-      expect(evaluatorDispatches()).toHaveLength(1);
-      expect(evaluatorDispatches()[0]?.payload.inputs).toEqual({
-        output: "yes",
-        expected_output: "yes",
-      });
-      const result = events.find((e) => e.type === "evaluator_result");
-      expect(result?.result.status).toBe("processed");
     });
   });
 });
